requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
httpx==0.25.2
python-multipart==0.0.6

# Required for Gemini API integration
# Set GOOGLE_GEMINI_API_KEY environment variable with your Google Gemini API key
